requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via pytest
click-log==0.3.2
click==6.7
coverage==4.5.1           # via pytest-cov
factory-boy==2.11.1
faker==0.8.16             # via factory-boy
flake8-isort==2.5
flake8-polyfill==1.0.2    # via pep8-naming
flake8-print==3.1.0
flake8-todo==0.7
flake8==3.3.0
inflection==0.3.1         # via pytest-factoryboy
isort==4.3.4              # via flake8-isort
mccabe==0.6.1             # via flake8
more-itertools==4.2.0     # via pytest
pep8-naming==0.7.0
pluggy==0.6.0             # via pytest
py==1.5.4                 # via pytest
pycodestyle==2.3.1        # via flake8, flake8-print, flake8-todo
pyflakes==1.5.0           # via flake8
pytest-cov==2.5.1
pytest-factoryboy==2.0.1
pytest-mock==1.10.0
pytest-sugar==0.9.1
pytest==3.6.2
python-dateutil==2.7.3    # via faker
pyyaml==3.12
six==1.11.0               # via faker, flake8-print, more-itertools, pytest, python-dateutil
termcolor==1.1.0          # via pytest-sugar
testfixtures==6.2.0       # via flake8-isort
text-unidecode==1.2       # via faker
